requirements.txt
channels==2.1.7
channels-redis==2.3.3
Django==1.11.20
pytest-asyncio==0.10.0
pytest-django==3.4.7
